backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
dj-rest-auth==6.0.0
django-allauth==0.63.6
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
django-filter==24.2
python-slugify==8.0.4
Pillow==10.3.0
dj-database-url==2.1.0
